move controller_voice.h
